+++ src/commands/__init__.py
@@ -6,3 +6,4 @@ from .cimg import CimgCommand  # noqa
 from .overlap import OverlapCommand  # noqa
 from .term import TermCommand  # noqa
 from .rulebook import RulebookCommand  # noqa
+from .help import HelpCommand  # noqa